change structure of servises
