+++ src/utils/index.js
@@ -3,11 +3,13 @@ const hashPassword = require('./hashPassword');
 const generatePassword = require('./generatePassword');
 const sendEmailWithPassword = require('./sendEmailWithPassword');
 const swagger = require('./swagger');
+const successMessage = require('./successMessage');
 
 module.exports = {
   ...generateToken,
   hashPassword,
   generatePassword,
   sendEmailWithPassword,
-  ...swagger
+  ...swagger,
+  successMessage
 };
