+++ src/utils/successMessage.js
@@ -1,0 +1,10 @@
+const statusCode = require('../statusCode');
+
+function successMessage(functionMessage) {
+  return {
+    code: statusCode.ok,
+    message: functionMessage,
+  };
+}
+
+module.exports = successMessage;
